move pipeline types to mongodb-support crate
