+++ crates/mongodb-agent-common/src/mongodb/mod.rs
@@ -1,17 +1,13 @@
-mod accumulator;
 mod collection;
 mod database;
-mod pipeline;
 pub mod sanitize;
 mod selection;
-mod stage;
 
 #[cfg(test)]
 pub mod test_helpers;
 
 pub use self::{
-    accumulator::Accumulator, collection::CollectionTrait, database::DatabaseTrait,
-    pipeline::Pipeline, selection::Selection, stage::Stage,
+    collection::CollectionTrait, database::DatabaseTrait, selection::selection_from_query_request,
 };
 
 // MockCollectionTrait is generated by automock when the test flag is active.
