+++ crates/mongodb-support/src/aggregate/mod.rs
@@ -1,0 +1,9 @@
+mod accumulator;
+mod pipeline;
+mod selection;
+mod stage;
+
+pub use self::accumulator::Accumulator;
+pub use self::pipeline::Pipeline;
+pub use self::selection::Selection;
+pub use self::stage::Stage;
